Parser for Noms type language.
Generated using pigeon (https://github.com/PuerkitoBio/pigeon).  This
patch introduces the 'parse' package, which provides an API for
parsing the Noms type language into a datastructure called
parse.Package. A Package contains a lookup table of defined types, a
list of container specializations it requires, and a set of imports
that are required to be satisfied in order for the package to be used.
No type-checking is currently performed; that'll be in a follow-on patch.

Towards Issue #281