Fix csv.ReportValidFieldTypes
csv.ReportValidFieldTypes() was always reporting that all rows could
be represented by any kind, no matter what the data. This is because
the schemaOptions data structure created in that function was being
updated inside a goroutine for which the main line of execution was
never waiting. All the CSV data would be read, each row sent over a
channel, and then the function would return as soon as it was done
reading.

Additionally, some 'go verify' cleanup in read.go and schema.go.
There's currently no need to expose schemaOptions or typeCanFit
outside the csv package, so these are made private.