Add tests for modifying chunked sets/maps/lists read from a chunkstore.
This is opposed to chunked sets/maps/lists that were constructed. The
difference is that constructed they will have their chunk values stored
in memory, whereas from a chunkstore they only have refs.

There was a bug fixed in 679305ab5bc63dc52facdcd8fe34973b467dbc91 which
caused this to crash. This patch adds tests and an extra assertion.