JS: Cache values in DataStore
The DataStore constructor now takes a second optional parameter
which is the number of bytes of values to cache.

DataStore is no longer a ChunkStore so it no longer has getRoot,
updateRoot, get, put or close.

Since DataStore now does its own caching CacheStore is no longer
useful and is now removed.

Fixes #1096