JS: Dynamically create a class per struct
The decoder now generates a class for a struct. Given:

```noms
struct MyStruct {
  x: Int8
  s: optional String
}
```

the generated class looks something like this:

```js
class MyStruct extends Struct {
  get x(): int8 { ... }
  setX(value: int8): MyStruct { ... }
  get s(): ?string { ... }
  setS(value: ?string): MyStruct { ... }
}
```

The base class `Struct` has nothing on it (except chunks). It
usefull in cases where a generic struct is wanted.

To reflect on Structs we introduce StructMirror.

This also means that Commit is now a real ~~boy~~type.

This does not yet handle Defs.

Towards #1157