Add clients/url_fetch_go
This adds a simple client which fetches a URL into a noms blob. In a separate change, I will change csv_importer to link up to this rather than reading from a file.

Note: I was sad to see that chunking performance here is quite slow. We can chunk at the rate of about 150kB second on a single core, and it's cpu-bound.