Go: Stop treating a chunk's hash as a hint for itself (#1758)
At some point, the Go ValueStore code started caching the hash of a
Chunk as a validation 'hint' for the Chunk itself.  JS never did
this. This must have addressed some edge case back when it was fatal
for validation code to run across a Chunk it hadn't seen and didn't
have a hint for. The downside is that doing this can cause us to send
a hint for every novel Chunk present in a writeValue payload in the
worst case.

Since I can't remember the edge case, and that edge case will no
longer be fatal anyway, removing this to avoid the (potentially
terrible) downside makes sense.