Make UTF-8 encode/decode write directly to a buffer (#2566)
We used to use a third party module that allocated new buffers all
the time. For Noms, we already have a buffer that we want to write
to.

For the rolling hash we use a scratch buffer that we reuse between
calls.

This also removes the DataView objects. Our usage is so basic that
having a DataView object wrapper is just overkill.

Fixes #2304