Noms type to GraphQL input types (#3263)
* Noms type to GraphQL input types

We already had a way to generate GraphQL output types from Noms types
but no way to do input types. Input types are used in arguments
position.

We currently do not do any conversion from the JSON like data in
argument position. We leave it to the users to determine if they want
to convert the data into noms data structures or go to Go structures
or operate directly on the JSON like data.

* Introduce argToNomsValue

Using `marshal.Marshal` is not going to work because structs are done as
`map[string]interface{}` in GraphQL input types.

Instead introduce argToNomsValue which correctly handles
GraphQL -> Noms correctly even for structs.