error changes (#7)
The existing noms codebase uses panics for signaling errors in most cases.  This is not idiomatic go, and it is not the optimal behavior when building cloud services, or products on top of noms as a panic in any go routine causes the entire process to exit.

This is a large step in removing panics, and bubbling errors up through the code.