go/store/types/edits: Changes to ensure AsyncSortedEdits always returned edits in stable order based on how they were added.
The way the SQL layer issues UPDATE statements to the storage layer, combined
with the change which introduced table_edit_accumulator meant that a Map Delete
followed by a Map Insert at certain batch sizes were being sent to an edit
accumulator and then fetched back out of it. When those come back from the edit
accumulator, they need to come back in stable order, so that the update doesn't
turn into a set followed by a delete instead.

AsyncSortedEdits did not currently make attempts to keep these edits in stable
order. In some places it's explicitly backwards (sorted_edit_itr for example).

This fixes it so that it is always stable.

Potentially regresses a memory utilization and sort size smoothing optimization
based around pairing small buffers with large buffers early in the process.

Also includes a fix to table_edit_accumulator to correctly order its
currently-in-progress commited edit provider when it is doing its stable merge.