go/cmd/dolt: commands/sqlserver: Add configuration for disabling MULTI_STATEMENTS support on the server.
Currently sql-server parses some MULTI_STATEMENTS incorrectly, in particular
statements that include substatements get split up incorrectly. This adds
server configuration for the yaml config, so that multi-statement handling can
be disabled server-side. This will allow things like CREATE TRIGGERs submitted
with a client that advertises MULTI_STATEMENTS support to work correctly with
Dolt sql-server. However, if the client actually sends a multi-statement in
this mode, even two SELECTs separated by a ';', for example, while the server
is running in this mode, it will not work.

To enable, place:

```yaml
behavior:
  disable_client_multi_statements: false
```

in the config.yaml for the server.