Add dolt_conflict_id to dolt_conflicts_$table_name
This PR makes manipulating the dolt_conflicts_$table_name tables easier.
It adds a new `dolt_conflict_id` column. The column is a hash of the primary keys for a
keyed table and the value hash of the row for a keyless table.

You can use the new column along with the `from_root_ish` column to select singular rows.
The `dolt_conflict_id` column does not uniquely identify conflict rows. Conflicts can
be aggregated across merges, which can result in two rows having the same `dolt_conflict_id`.