Rewrite broken-merge.bats in enginetest package
Rewrite of broken-merge.bats in engine test package. Includes `pmd`, which is helper function
that parses "merge documents".

We expect a high number of merge tests to be created and it would help if writing
those tests became less repetitive and easy to understand. The `pmd` function helps us do that.

I looked into using //go:generate tags as a code generation solution, but it didn't pan out.