go: sqle: cluster: Implement a first pass at dolt_cluster_ack_writes_timeout_secs.
Setting this system variable to a non-zero value on a primary replica in a
sql-server cluster will cause dolt to block a SQL client performing a commit
until that client's commit is fully replicated to the replicas.

If there is a timeout, currently a warning is logged in the logs.