go/store/nbs: table_reader.go: Optimize hasMany to use a binary search instead of a linear search.
With the addition of `errorIfDangling`, hasMany is on the critical path for
COMMIT and dataset HEAD operations. The other places its heavily used is GC and
push/pull. The linear scan can be the right tradeoff for small table files and
for large table files with a lot of addresses to look for, but it's a bad
tradeoff for most errorIfDangling checks in particular, especially on large
databases with large table files.

For now we will always use the binary search, but a heuristic based approach
which could pick between the two would be better.