dolt table import: json,csv: Support BOM file headers.
The semantics are as follows:

For CSV files, the default import is an uninterpreted character encoding where
newline has to match 0xa and the delimeters have to match. In general Dolt
expects UTF8, but non-UTF8 characters in string fields can make it through to
the imported table for encodings which are close enough to ASCII, for example.
If there is a UTF8, UTF16LE or UTF16BE BOM header, then character decoding of
the input stream switches to the indicated encoding.

For JSON files, the default import is UTF8 character encoding. If there is a
UTF8, UTF16LE or UTF16BE BOM header, then character decoding of the input
stream switches to the indicated encoding.