go: doltcore/remotestorage: reliable: Fix leaked goroutines when the context is canceled while reading the HTTP response.
io.Copy into an io.PipeWriter will block until all the bytes have been
delivered or the reader is closed. reliable/http StreamingResponse was
constructed to only cancel the request context on Close(), not also clear the
Reader. The Reader should also be closed to ensure all finalization can still
happen if the Write to the PipeWriter is currently blocked when the context is
canceled.