go/store/{chunks,nbs}: Move to a model where MarkAndSweepChunks returns a finalizer.
Previously MarkAndSweepChunks would finalize itself by either appending the
table table files to the store or by swaping to the table files. Now it returns
a GCFinalizer instance which is used by the caller to effect the desired
change. This is necessary for implementing Full GC, where we actually want to
do both with the same GC result, but at different phases of the GC process.

This commit leaves the HasManyF return from NomsBlockStore.MarkAndSweepChunks
as a TODO.