libraries/doltcore/remotestorage: Improve connection reuse when fetching chunks from remote storage.
Improves performance of fetches from DoltHub, sql-server, and doltlab.

Improves some situations where routers and network access points do not respond
well to clients which open a great number of connections to the same TCP
endpoint, many of which live for a very short period of time.

http.DefaultClient comes with a default MaxIdleConnsPerHost of 2. Fetching with
a high number of concurrent downloads results in many connections not being
reused efficiently because they cannot get back into the idle pool because it
is so small. Here we increase the MaxIdleConnsPerHost to be large enough so
that our active connections during a fetch will fit within it.

Note: S3 limits a single HTTP connection to 100 requests, after which it is
closed server-side. So currently dolt can still require a large number of new
connections over the course of a pull, and the rate at which they are opened
will depend on a number of factors including available throughput, round trip
time resolving storage locations, etc. But this change will always be a big
improvement over the old behavior.