[no-release-notes] go: binlogreplication: tests: Parallelize the test suite, slightly improve process management.
Move the global state to a harness struct, and change all interactions to go
through the harness, so that we can run these tests in parallel.

We also clean up the process management a bit, since it was subject to leaving
straggling sql-server and mysqld instances for me, usually because I would ^C
it when it took too long after I ran `go test ./...`. The two major changes
are:

1) In the happy case, we now gracefully terminate the child and assert that it
exited success.

2) We now explicitly try to get helpful signals to all children when the test
binary itself receives a SIGINT.