[no-release-notes] go: sqle: dolt_gc: Purge read caches associated with the DoltDB as we begin GC.
The right place to do this is in the safepoint controller, at BeginGC. We need
to do it after keeperFunc is installed on the NomsBlockStore, so that all read
dependencies are taken. We also need to do it before we register all sessions
whose cached state we need to visit.