go: remotestorage: chunk_store.go: Clean up ChunkCache.
When DoltChunkStore was implemented, it needed to do write buffering in order
to read its own writes and to flush table files to the upstream. At some point,
read caching and caching of has many results was added onto the write buffer,
creating confusion and despair. This change separates back out the use cases.