go: cmd/dolt: Move to closing the CliContext, instead of closing the response from Queryist QueryEngine().
This makes the lifecycle of the LateBinder resources more correct, since they
should typically live until the end of the process. In test contexts, in
particular, it was common to have one command see the closer != nil, close it,
and still receive the cached Queryist from the CliContext on later Command.Exec
invocations.