[no-release-notes] go/utils/publishrelease/buildpgobinaries.sh: Update release build script to use ICU4C-enabled toolchains.
We install the statically built ICU4C packages into the sysroots before building.

We pass icu_static as a build tag.

We add CXX support to the Cgo build.

We bump build image to Debian trixie (newer MinGW is necessary) and we pick up cross tools which have built for that base image.

We add -static-libgcc and -static-libstdc++ to the Windows build, as necessary.