go: libraries/doltcore/dbfactory: Do not use EventsUnaryClientInterceptor when interacting with remotestorage.
This events source is too detailed and emits too many events to be appropriate
for our usage metrics. It is useful in an operational context for DoltHub and
Hosted, but it should be used by default in Dolt itself. We will revisit this
events source if we need it for internal infrastructure, but for now we should
disable it.