go: sqle/dsess: autoincrement_tracker.go: Fix race condition in initialization.
When initializing the autoincrement_tracker, we look at the current value of the autoincrement sequence on the table across every branch. We do this concurrently, and take the highest value we find as the initial value to use across the `dolt` run.

This fixes a race condition which would cause us to calculate the wrong highest value across all branches.