go: store/datas/pull: pull_chunk_tracker.go: Optimize memory use when backing up to an AWS remote.
PullChunkTracker is responsible for making the HasMany calls against the
destination and batching up absent hashes into HashSets which will be delivered
to GetManyCompressed and eventually written into table files which are
uploaded. This code is used for both pull and push, when the destination is the
"local" database or when destination is the remote database respectively. It is
used when the remote is both doltremoteapi, thus every HasMany call is an RPC,
and when the remote is something like file:// or aws://, thus the table file
indexes for the remote are in memory and HasMany calls are very quick.

Different operational characteristics of the various dependencies mean that
sometimes a Pull is prone to build up large sets of hashes waiting for HasMany
calls, whereas other times it is prone to build up large sets of absent hashes
which are waiting for the fetcher thread(s) to take them.

Previously, PullChunkTracker was structured to accumulate HasMany responses and
wait to batch them into appropriately-sized batches for GetManyCompressed until
the fetcher threads asked for them. This meant that if HasMany batches were
very small, because HasMany was very fast, we would accumulate a large number
of very small HashSets. These HashSets would take up large amounts of memory.
Accumulating the batches as the HasMany responses come in is more memory
efficient and should be no slower - we will always accumulate the full batches,
and in basically the same order.

Tested by pushing a large database to an AWS remote and memory profiling the
result.